move classes to proper packages
