+++ graylog2-server/src/main/java/org/graylog2/bindings/ServerBindings.java
@@ -59,7 +59,7 @@ import org.graylog2.inputs.InputRuntimeStatusProvider;
 import org.graylog2.inputs.InputStateListener;
 import org.graylog2.inputs.PersistedInputsImpl;
 import org.graylog2.lookup.LookupModule;
-import org.graylog2.metrics.EntityMetricsModule;
+import org.graylog2.metrics.entity.EntityMetricsModule;
 import org.graylog2.plugin.cluster.ClusterConfigService;
 import org.graylog2.plugin.cluster.ClusterIdFactory;
 import org.graylog2.plugin.cluster.ClusterIdService;
