+++ graylog2-server/src/main/java/org/graylog2/commands/Server.java
@@ -33,7 +33,6 @@ import org.graylog.collectors.CollectorsModule;
 import org.graylog.enterprise.EnterpriseModule;
 import org.graylog.events.EventsModule;
 import org.graylog.events.processor.EventDefinitionConfiguration;
-import org.graylog2.metrics.cache.MetricsCacheConfiguration;
 import org.graylog.grn.GRNTypesModule;
 import org.graylog.mcp.server.McpServerModule;
 import org.graylog.metrics.prometheus.PrometheusExporterConfiguration;
@@ -98,6 +97,7 @@ import org.graylog2.indexer.rotation.RotationStrategyBindings;
 import org.graylog2.inputs.transports.NettyTransportConfiguration;
 import org.graylog2.lookup.adapters.dnslookup.DnsLookupAdapterConfiguration;
 import org.graylog2.messageprocessors.MessageProcessorModule;
+import org.graylog2.metrics.entity.cache.MetricsCacheConfiguration;
 import org.graylog2.migrations.MigrationsModule;
 import org.graylog2.notifications.Notification;
 import org.graylog2.notifications.NotificationService;
