+++ graylog2-server/src/main/java/org/graylog2/metrics/entity/EntityMetricValues.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.metrics;
+package org.graylog2.metrics.entity;
 
 import java.util.Collections;
 import java.util.HashMap;
