+++ graylog2-server/src/main/java/org/graylog2/metrics/entity/cache/CacheResult.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.metrics.cache;
+package org.graylog2.metrics.entity.cache;
 
 import java.util.Collection;
 import java.util.HashMap;
